+++ src/Pokemon.java
@@ -1,0 +1,81 @@
+import java.util.Arrays;
+import java.util.List;
+
+public abstract class Pokemon {
+    private String type;
+    private String name;
+
+    int level;
+    int hp;
+    private String food;
+    private String sound;
+
+    private List<String> attacks;
+
+    public Pokemon(String name, int level, int hp, String food, String sound) {
+        this.name = name;
+        this.level = level;
+        this.hp = hp;
+        this.food = food;
+        this.sound = sound;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getLevel() {
+        return level;
+    }
+
+    public void setLevel(int level) {
+        this.level = level;
+    }
+
+    public int getHp() {
+        return hp;
+    }
+
+    public void setHp(int hp) {
+        this.hp = hp;
+    }
+
+    public String getFood() {
+        return food;
+    }
+
+    public void setFood(String food) {
+        this.food = food;
+    }
+
+    public String getSound() {
+        return sound;
+    }
+
+    public void setSound(String sound) {
+        this.sound = sound;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public List<String> getAttacks() {
+        return attacks;
+    }
+
+    public void setAttacks(List<String> attacks) {
+        this.attacks = attacks;
+    }
+
+    public void getpokemonList();
+}
+
