+++ src/PokemonGym.java
@@ -1,0 +1,2 @@
+interface PokemonGym {
+}
