+++ src/PokemonGymOwner.java
@@ -1,0 +1,17 @@
+public class PokemonGymOwner extends PokemonTrainer {
+    private String name;
+
+    public PokemonGymOwner(String name) {
+        super();
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
