+++ src/Variabelen.java
@@ -1,11 +1,6 @@
-import java.util.Arrays;
-import java.util.List;
-
 public class Variabelen {
     /*deze komen meerdere keren voor*/
-    type; /*""*/
     List<String> attacks = Arrays.asList(/*voer de aanvallen hier in*/);
-    name;/*""*/
 
     /*deze variabelen komen eenmaal voor*/
     charizard;/*🔥*/
@@ -15,11 +10,7 @@ public class Variabelen {
     raichu;/*⚡*/
     gyarados;/*🌊*/
     List<Pokemon> pokemons;
-    int level;
-    int hp;
-    food;/*""*/
-    sound;/*""*/
-    
+
     /*deze variabele is optioneel. Je hoeft deze niet te definieren, maar als je hem definieert, dan moet je hem ook ergens gebruiken*/
     village;/*""*/
